perf(clouddriver): Remove down instance check in disable server group stage.
When a server group is disabled in AWS, you don't need to check for down instances. That's not the point of the stage, as it does not scale down the instances. It simply detaches them from the load balancer so they no longer receive traffic. This step alone is vastly slowing down our release cadence.